Start of adding build tools
Two new files added in this commit meant to aide with building the
project.

- build.sh
Currently this just execs square-trim.sh, the other new file in this
commit, but in the future this should idealy do everything required to
set up the static site when executed

- square-trim.sh
This file runs over a directory, trimming and squaring all .png images
within.  It is currently used to make nice, small, square images of the
relic icons.  It depends on imagemagick (currently only tested with IM
7)